buildkitd: allow multiple units for gc config
In addition to parsing the raw number of bytes, we can additionally
supporting reading short notations from strings, such as "50MB", or
"10GB". We can also support different percentages, such as "20%", which
allows for consuming a maximum percentage of the disk easily
(previously, only the default was set to 10%, with no ability to change
to use an arbitrary percentage).

Signed-off-by: Justin Chadwell <me@jedevc.com>